Fix for issue 2613
The upgrader is breaking replication (and not just dsreplication).

The problem occurs because the upgrader is systematically overwriting the contents of the admin-backend.ldif file when it should not.  The fix consists on not updating this file unless it does not exist.

The same fix has been applied to tool.properties since it was also systematically overwritten (and so the user lost the changes made to it).

Appart some fixes related to message formatting are also included in the attached diff file.  The code used to display messages with HTML formatting in CLI mode, so specific messages have been added when the upgrade is run from the command line.

Finally the code of the upgrader has been updated to use the MenuBuilder class.  This is done to have a consistent output with the rest of the command-lines. 


git-svn-id: https://svn.forgerock.org/opendj/trunk@3485 41b1ffd8-f28e-4786-ab96-9950f0a78031