Fix 2319 - protect acces to dc=replicationChanges by default Description dc=replicationChanges is enabled by default on all the replication servers and show all the modifications done on all replicated suffixes. It may therefore contain sensitive information and should be available only administrators.
Fix
This is fixed by a global ACI on the "dc=replicationChanges" suffix denying all ops for all users. Only "directory manager" can access to the changes by default or change the ACI.
Unit test written and passed successfully individually but set disabled because it fails when ran as part of the whole suite (issue 1569 to be addressed asap).



git-svn-id: https://svn.forgerock.org/opendj/trunk@3511 41b1ffd8-f28e-4786-ab96-9950f0a78031