Fix for  issue 2587 : Schema initialisation fails with dependency error
dsreplication enable fails to initialise schema of the peer if the user schema
contains a new objectClass dependent on a new attributeType.

This was happening because the validation of the new objectClasses was done using the
old schema while it must be done with the new schema currently being modified. 

git-svn-id: https://svn.forgerock.org/opendj/trunk@3532 41b1ffd8-f28e-4786-ab96-9950f0a78031