This fixes 2724 : and enable/fix some more replication tests.
If the Replication Server is stopped while the ServerHandler writer threads are blocked waiting for some window credit they are blocked and don't stop
immediately.

This issue is minimal for real life but can cause unit test failure because replication Servers are stopped and started a number of times during the tests.

This commit also enabled some more replication tests and change them to make the
order in which they are run more predictible.

git-svn-id: https://svn.forgerock.org/opendj/trunk@3542 41b1ffd8-f28e-4786-ab96-9950f0a78031