Fix for issue 2612.
Initialization of the ADS fails when both servers where configured for replication separately.

The problem is that the entry that we read from the ADS is not complete (is missing the public key attribute).  So when we use the server entry we read from one ADS to register that server in the other ADS the registration is not complete.

The fix consists simply on reading the public key certificate in the ADS when we construct the object representing the server.


git-svn-id: https://svn.forgerock.org/opendj/trunk@3555 41b1ffd8-f28e-4786-ab96-9950f0a78031