Fix for issue 2446 part #1.
Add support to the admin framework for specifying one or more "default" managed objects which should be created automatically when a parent managed object is created. The XML mark-up is provided as part of the parent managed object's relation definition as follows (the sample is taken from the LocalDBBackend XML definition):

     <adm:relation name="local-db-index">
       <adm:one-to-many naming-property="attribute">

         <adm:default-managed-object name="aci">
           <adm:property name="index-type">
             <adm:value>presence</adm:value>
           </adm:property>
           <adm:property name="attribute">
             <adm:value>aci</adm:value>
           </adm:property>
         </adm:default-managed-object>

         <adm:default-managed-object name="entryUUID">
           <adm:property name="index-type">
             <adm:value>equality</adm:value>
           </adm:property>
           <adm:property name="attribute">
             <adm:value>entryUUID</adm:value>
           </adm:property>
         </adm:default-managed-object>

         <adm:default-managed-object name="objectClass">
           <adm:property name="index-type">
             <adm:value>equality</adm:value>
           </adm:property>
           <adm:property name="attribute">
             <adm:value>objectClass</adm:value>
           </adm:property>
         </adm:default-managed-object>
       </adm:one-to-many>

The functionality is implemented on the client-side API in a generic manner so that it is independent of the underlying driver implementation (so the default indexes would be created if we had an offline direct to file driver as well).



git-svn-id: https://svn.forgerock.org/opendj/trunk@3585 41b1ffd8-f28e-4786-ab96-9950f0a78031