Fix for issue 2791: dsreplication enable command assumes Administrator doesn't already exist
The fix consists of checking whether the specified administrator exists or not before creating it (this is done even in the case where no server had been registered).

In addition to that a bug in the way the server registry comparison is made has been fixed.  Before these changes when we tried to enable a suffix between two servers that were already replicated, dsreplication enable failed.


git-svn-id: https://svn.forgerock.org/opendj/trunk@3622 41b1ffd8-f28e-4786-ab96-9950f0a78031