Fix issue 2827: Messages constructed from a MessageBuilder should have a sensible category and severity.
Modify the MessageBuilder.toMessage() method so that constructed Messages have a sensible Category and Severity. Now the Category and Severity are based on the first object appended to the MessageBuilder. If that object is a Message then the MessageBuilder will use the Category and Severity of that Message, otherwise it will continue the current default behavior which is to use Category.USER_DEFINED and Severity.INFORMATION.

git-svn-id: https://svn.forgerock.org/opendj/trunk@3673 41b1ffd8-f28e-4786-ab96-9950f0a78031