Fix for issue 2818 (Cannot start replication ifthe second server created using quicksetup tool)
I have not been able to reproduce the problem as described in the bug report but there was a bug with the ApplicationTrustManager used by the Web Start installer: in fact when the user accepted the certificates it was not updated properly and this generated connection problems.

Since with these modifications I am not able to reproduce the issue I am marking the bug as fixed.


git-svn-id: https://svn.forgerock.org/opendj/trunk@3677 41b1ffd8-f28e-4786-ab96-9950f0a78031