Fix for issue 2830: Improvements to build "test" target.
This change modifies the Ant "test" target so that it is as lightweight as possible but forgiving for first-time users.

The test target now checks to see if there is an OpenDS.zip package in build/packages and, if there is, it simply recompiles the Java source files (no message generation or admin framework code generation). If no OpenDS.zip package is found then it automatically invokes the "package" target so it is now possible to do the following:

   ./build.sh clean
   ./build.sh test

git-svn-id: https://svn.forgerock.org/opendj/trunk@3680 41b1ffd8-f28e-4786-ab96-9950f0a78031