Fix for issue 2862 (Online backup command exit status 0 while backup is failed).
The code used to consider the errors during backup as warnings.  In general the errors during the backupmean that at least one backend could not be backed up, the modifications simply update the code to return TaskState.STOPPED_BY_ERROR instead of TaskState.COMPLETED_WITH_ERRORS.  This is more accurate because if one of the backends is not backed up we should return an error code and because every error detected in the backup code stops the backup processing of a particular backend.


git-svn-id: https://svn.forgerock.org/opendj/trunk@3715 41b1ffd8-f28e-4786-ab96-9950f0a78031