Fix for 2312.
Currently the lines are only wrapped when there is an space. If not it
has been decided to make the cell wider than what has been set in the
code. I guess that the reason for doing this is to avoid problems
grepping on whole words, however I think that the side effect is more
negative than the problem that we try to solve.

In general tables are printed to help humans to understand the output
(and not for scripts). Breaking the whole table structure goes against
this principle. 


git-svn-id: https://svn.forgerock.org/opendj/trunk@3719 41b1ffd8-f28e-4786-ab96-9950f0a78031