The Replication Server thread that is reading changes from the database to propagate them to the Directory Servers was sometimes leaving some Database cursors open.
This was leading to Deadlock Exceptions that was causing shutdown of
the Replication Server

When the 2 Replication Servers used in this test were dead the
Replication was dead.

These changes fix the code so that we always release the cursors
and catch the Deadlock Exception in case we would meet transient
Deadlock even though we believe it's not possible with the current
code.


git-svn-id: https://svn.forgerock.org/opendj/trunk@3810 41b1ffd8-f28e-4786-ab96-9950f0a78031