Fix for issue 2910 (uninstall : javaexception when specifying bad admin dn)
The fix consists of checking whether the uid and pwd arguments are null or not.

Appart from this a bug has been fixed in the uninstall parser: the method parseArguments was called twice (once in UninstallLauncher and another in UninstallCliHelper).  This made that providing any non-boolean argument in the command-line resulted in a error.


git-svn-id: https://svn.forgerock.org/opendj/trunk@3833 41b1ffd8-f28e-4786-ab96-9950f0a78031