Fix some font formatting issues in the setup.
Try to minimize the cases where issue 2915 can affect to the setup by trying to reconnect more often to the server after it is started.

Since the start can take long, use the same point mechanism to show progress in non verbose mode we use for the import.  This way the user does not have the impression that we are stuck on the start.


git-svn-id: https://svn.forgerock.org/opendj/trunk@3840 41b1ffd8-f28e-4786-ab96-9950f0a78031