Fix for issue 2919 (translation of version related properties causes error)
The issue comes because we are parsing the output of a command-line and this output is localized.  Since the option we are using is not part of the public interface.   This information can be used not only by the BuildInformation class but also by the support team so it is better not to localize it.  The fix consists of removing these properties from the properties file and define some String constants in SetupUtils so that can be shared by BuildInformation and DirectoryServer classes.



git-svn-id: https://svn.forgerock.org/opendj/trunk@3859 41b1ffd8-f28e-4786-ab96-9950f0a78031