Fix for issue 2886 (Cannot get status from running server without authenticating)
The code has been modified (with no new arguments) in order to work in a best-effort mode.  If the user does not provide authentication we will display the same information independently of whether the server is running or not.  We still display the messages informing that authentication is required to display some monitoring data.

The status panel code has also been updated to have the same behavior.

However it will be required anyhow to provide authentication to display monitoring information since this information is available only through LDAP.

In addition to this changes the following issues have been fixed:

1. Close the ManagementContext created by the LDAPConsoleInteraction object so that we have only one open connection to the server.
2. Fix a bug when the user chose to accept automatically certificates (the default trust manager was used in this case and we could not connect).


git-svn-id: https://svn.forgerock.org/opendj/trunk@3893 41b1ffd8-f28e-4786-ab96-9950f0a78031