Fix for issue 2969 (In some cases the arguments set for offline operations (import-ldif, export-ldif) are the ones of the online mode)
The problem comes because when we are dealing with scripts that can have both online and off-line modes we call _script-util to know whether we must run the script in online or off-line modes.  This makes the environments variables to be set, so when we call the second time _script-util (to set the environment for the off-line mode) nothing is changed if the user specified not to overwrite the environment variables.

The fix consists on resetting the environment variables to its original state before calling _script-util for the second time.


git-svn-id: https://svn.forgerock.org/opendj/trunk@3895 41b1ffd8-f28e-4786-ab96-9950f0a78031