Fix for issue 2972 (The usage of stop-ds.bat shows the class name)
The fix consists on updating the scripts and the java code to reflect correctly the name of the script and be consistent with all the other batch files (we don not include the .bat on the usage).

Apart from that the modifications include a modification in StatusPanelLauncher so that we use the ArgumentParser class in the whole class (this avoids duplicated code and fixes a bug with the usage).


git-svn-id: https://svn.forgerock.org/opendj/trunk@3900 41b1ffd8-f28e-4786-ab96-9950f0a78031