Fix for issue 3048 (Missing "Are you sure ?" confirmation question in setup --cli)
In interactive mode display a summary of the parameters provided by the user (as the one that we have in the GUI).  If the user is not satisfied with the provided parameters we propose to revisit all of them again (the default values that we will propose are the ones the user provided).


git-svn-id: https://svn.forgerock.org/opendj/trunk@3954 41b1ffd8-f28e-4786-ab96-9950f0a78031