Fix for issue 3120 (Bad behavior of the status CLI and status panel if 2 LDAP connection handlers)
The issue is not related to the fact that we have several LDAP connection handlers but to the fact that the default listen address (0.0.0.0) is not written to cn=config nor the value is returned when asking for it explicitly through ldapsearch.  The fix consists of using the default value if the explicit value of the listen address is not found.


git-svn-id: https://svn.forgerock.org/opendj/trunk@4105 41b1ffd8-f28e-4786-ab96-9950f0a78031