Fix for 3338 (upgrade breaks the java_home parameter)
The problem is that the upgrade wipes out the set-java-home files.  This fix recreates them by calling dsjavaproperties script if they did exist before the upgrade is run.  The approach of generating the files is more appropriate than keeping the old files since they might be changes in the scripts after an upgrade.



git-svn-id: https://svn.forgerock.org/opendj/trunk@4344 41b1ffd8-f28e-4786-ab96-9950f0a78031