Fix for issue 3477 : OpenDS runs out of synchronization or crashes
This problem happens because the cache of messages kept by the
Replication Servers are limited only by the a number of messages.

In the particular case of this tests, the entries used are very large
and modifications are done using large attribute values.
This cause the Replication Server caches to become too large and
the JVM runs out of memory.

The solution is to account not only for the number of messages but
also for the total number of bytes in the queues and to stop
caching messages in the queue when the maximum size is reached.


git-svn-id: https://svn.forgerock.org/opendj/trunk@4488 41b1ffd8-f28e-4786-ab96-9950f0a78031