Fix for issue 3554 (Control Panel : Error displayed after deleting several entries).
The problem came because we were setting a subtree delete control on the same di
r context used to refresh the tree.  This generate some problems with the connec
tion that got closed.  The fix consists on analyzing the node that is being dele
ted, if it contains over 40 subordinates, we use another connection and the subt
ree delete control, if not we delete the subtree recursively.

The number of notifications in the logs and the calls to the event thread have b
een reduced since they had quite a bit of impact in performance.

The code has been updated so that after restarting the server we try to reuse th
e previously provided credentials, instead of simply asking the user to login ag
ain.


git-svn-id: https://svn.forgerock.org/opendj/trunk@4576 41b1ffd8-f28e-4786-ab96-9950f0a78031