Fix for issue 3566 (Increase the timeout of the windows service and consider making it configurable).
The fix increases the number of tries used to figure out if the server is started or not.  This is required in particular after reboot since the system can be really loaded.

The user can modify the default tries value (100 tries with an interval of 5 seconds between tries) by setting the value of the OPENDS_WINDOWS_SERVICE_START_NTRIES  environment variable (it must be a SYSTEM environment variable in order to be taken into account).

git-svn-id: https://svn.forgerock.org/opendj/trunk@4592 41b1ffd8-f28e-4786-ab96-9950f0a78031