Fix for issue 3613 (control panel : incorrect output messages when deleting subtree)
The difference on the method used to delete entries (with or without control) is not related to whether the delete is executed using the popup menu or the button.  It is related to the number of entries below the entry that has been selected to be deleted.  The threshold used currently is of 40 entries; since the common case is to delete subtrees, the threshold has been changed to 1.  So when an entry is selected and contains entries below it, the control will be used systematically (instead of a recursive delete).


git-svn-id: https://svn.forgerock.org/opendj/trunk@4639 41b1ffd8-f28e-4786-ab96-9950f0a78031